+++ src/Util/Player/LivePlayerDataBuilder.php
@@ -1,0 +1,68 @@
+<?php
+
+namespace srag\Plugins\Opencast\Util\Player;
+
+use xoctConf;
+use xoctRequest;
+
+/**
+ * Class LivePlayerDataBuilder
+ * @package srag\Plugins\Opencast\Util\Player
+ * @author  Theodor Truffer <tt@studer-raimann.ch>
+ */
+class LivePlayerDataBuilder extends PlayerDataBuilder
+{
+
+    public function buildStreamingData() : array
+    {
+        $episode_json = xoctRequest::root()->episodeJson($this->event->getIdentifier())->get(
+            [],
+            '',
+            xoctConf::getConfig(xoctConf::F_PRESENTATION_NODE)
+        );
+
+        $episode_data = json_decode($episode_json, true);
+        $media_package = $episode_data['search-results']['result']['mediapackage'];
+
+        $streams = [];
+        if( isset($media_package['media']['track'][0]))
+        {
+            foreach ($media_package['media']['track'] as $track) {
+                $role = strpos($track['type'], self::ROLE_MASTER) !== false ? self::ROLE_MASTER : self::ROLE_SLAVE;
+                $streams[$role] = [
+                    "content" => $role,
+                    "sources" => [
+                        "hls" => [
+                            [
+                                "src" => $track['url'],
+                                "mimetype" => $track['mimetype'],
+                                "isLiveStream" => true
+                            ]
+                        ]
+                    ]
+                ];
+            }
+        } else {
+            $track = $media_package['media']['track'];
+            $streams[] = [
+                "content" => self::ROLE_MASTER,
+                "sources" => [
+                    "hls" => [
+                        [
+                            "src" => $track['url'],
+                            "mimetype" => $track['mimetype'],
+                            "isLiveStream" => true
+                        ]
+                    ]
+                ]
+            ];
+        }
+
+        return [
+            "streams" => array_values($streams),
+            "metadata" => [
+                "title" => $this->event->getTitle(),
+            ],
+        ];
+    }
+}