+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Improvement: show column for unprotected link only if such link is present (and restricted to edit_videos permission)
 - Feature: config for presigning urls (possible performance boost)
 - Feature: config for synchronous loading of table (possible performance boost)
+- Fix: internal player showed two streams but no quality selector on a single stream with multiple qualities
 - Fix: omit error when emptying trash with old Opencast object via cron job
 
 ## [3.4.2]